image_project/requirements.txt
fastapi
uvicorn
pydantic
torch
diffusers
python-dotenv